This is the development branch for doing latency measurements on cpu and gpu, and creating an exporter to nnp and onnx:
New definition of the networks for zoph and random wired search space. Added functions to measure latency and to export to onnx for zoph and random wired and mobilenet search spaces. Bug fixing for calculation of latency. Two Latency estimators: LatencyEstimator(), a module-based latency estimator, LatencyGraphEstimator(), and a graph-based latency estimator. Both work for static as well as for dynamic modules. Central script that shows examples how to trigger the calculation of the latency and exporting to onnx for zoph, random wired and mobilenet search spaces. The latency estimation can be carried out on cpu or gpu. Another simpler script with an example for quick latency calculation. Updated Documentation- Code works with nnabla version 1.19.0.